feat(core): per-agent runOn execution backend + Provider.submitDagRun
Add Agent.runOn ('local' | 'temporal') — where an agent's runs orchestrate,
distinct from the LLM provider. Persisted in the versioned DAG (schema + store
round-trip). Promote SubmitDagRunOptions to the core Provider interface as an
optional submitDagRun, so the dashboard can route durable runs without importing
the temporal package; TemporalProvider now uses the core type.
