+++ packages/temporal-provider/src/index.ts
@@ -2,5 +2,4 @@ export * from './provider.js';
 export * from './worker.js';
 export { createTemporalSpawnNode, type CreateTemporalSpawnNodeOptions } from './node-spawn.js';
 export type { RunNodeActivityInput, RunDagActivityInput, RunDagActivityResult } from './activities.js';
-export type { SubmitDagRunOptions } from './provider.js';
 export type { RunAgentWorkflowInput, RunAgentWorkflowResult } from './workflows.js';
